Restructure Vorprüfung route folder structure
