+++ packages/dito/app/routes/vorpruefung.ergebnis.tsx
@@ -6,7 +6,7 @@ import { json, type LoaderFunctionArgs } from "@remix-run/node";
 import { useLoaderData } from "@remix-run/react";
 import { getAnswersFromCookie } from "cookies.server";
 import { precheck } from "resources/content";
-import { type Answers } from "./vorpruefung.$questionId/route";
+import { type Answers } from "./vorpruefung.$questionId";
 
 export async function loader({ request }: LoaderFunctionArgs) {
   const { answers } = await getAnswersFromCookie(request);
@@ -43,6 +43,7 @@ export default function Result() {
       },
     ],
   }));
+
   return (
     <>
       <Container>
